refactor(tools): migrate Slack tools to directory structure
- Split monolithic slack.ts (661 lines) into organized directory structure
- Created slack/types.ts with SlackToolsConfig and SlackClientConfig interfaces
- Created slack/auth.ts with authentication helpers using closures
- Split 4 tools into individual files in slack/tools/:
  - send-slack-message.ts
  - notify-slack.ts
  - get-slack-channels.ts
  - get-slack-messages.ts
- Created slack/index.ts with default exports and factory function
- Updated web/index.ts to export from new slack/index.js location
- Updated test imports (2 locations) to use new path
- All 23 tests passing
- Build successful
- No breaking changes to public API

Follows the pattern established by Confluence and web-search tools.
Part of Phase 2 of the tools refactoring plan.
