+++ packages/tools/src/web/index.ts
@@ -9,6 +9,6 @@ export * from './scraper.js';
 export * from './html-parser.js';
 export * from './url-validator.js';
 export * from './web-search/index.js';
-export * from './slack.js';
+export * from './slack/index.js';
 export * from './confluence/index.js';
